+++ wgpu-core/src/command/compute.rs
@@ -17,7 +17,10 @@ use crate::{
         memory_init::{fixup_discarded_surfaces, SurfacesInDiscardState},
         pass::{self, flush_bindings_helper},
         pass_base, pass_try,
-        query::{end_pipeline_statistics_query, validate_and_begin_pipeline_statistics_query},
+        query::{
+            end_pipeline_statistics_query, record_pass_timestamp_writes,
+            validate_and_begin_pipeline_statistics_query,
+        },
         ArcCommand, ArcPassTimestampWrites, BasePass, BindGroupStateChange, CommandEncoder,
         CommandEncoderError, DebugGroupError, EncoderStateError, InnerCommandEncoder, MapPassErr,
         PassErrorScope, PassStateError, PassTimestampWrites, QueryUseError, StateChange,
@@ -632,6 +635,8 @@ pub(super) fn encode_compute_pass(
                 indirect_draw_validation_resources: parent_state.indirect_draw_validation_resources,
                 snatch_guard: parent_state.snatch_guard,
                 debug_scope_depth: &mut debug_scope_depth,
+                query_set_writes: parent_state.query_set_writes,
+                deferred_query_set_resolves: parent_state.deferred_query_set_resolves,
             },
             binder: Binder::new(),
             temp_offsets: Vec::new(),
@@ -670,6 +675,8 @@ pub(super) fn encode_compute_pass(
         if let Some(tw) = timestamp_writes.take() {
             tw.query_set.same_device(device).map_pass_err(pass_scope)?;
 
+            record_pass_timestamp_writes(&tw, state.pass.base.query_set_writes);
+
             let query_set = state
                 .pass
                 .base
@@ -688,6 +695,9 @@ pub(super) fn encode_compute_pass(
                     .or(tw.end_of_pass_write_index)
                     .map(|i| i..i + 1)
             };
+            let raw_query_set = query_set
+                .try_raw(parent_state.snatch_guard)
+                .map_pass_err(pass_scope)?;
             // Range should always be Some, both values being None should lead to a validation error.
             // But no point in erroring over that nuance here!
             if let Some(range) = range {
@@ -696,12 +706,12 @@ pub(super) fn encode_compute_pass(
                         .pass
                         .base
                         .raw_encoder
-                        .reset_queries(query_set.raw(), range);
+                        .reset_queries(raw_query_set, range);
                 }
             }
 
             Some(hal::PassTimestampWrites {
-                query_set: query_set.raw(),
+                query_set: raw_query_set,
                 beginning_of_pass_write_index: tw.beginning_of_pass_write_index,
                 end_of_pass_write_index: tw.end_of_pass_write_index,
             })
@@ -812,13 +822,19 @@ pub(super) fn encode_compute_pass(
                     query_index,
                     None,
                     &mut state.active_query,
+                    state.pass.base.snatch_guard,
                 )
                 .map_pass_err(scope)?;
             }
             ArcComputeCommand::EndPipelineStatisticsQuery => {
                 let scope = PassErrorScope::EndPipelineStatisticsQuery;
-                end_pipeline_statistics_query(state.pass.base.raw_encoder, &mut state.active_query)
-                    .map_pass_err(scope)?;
+                end_pipeline_statistics_query(
+                    state.pass.base.raw_encoder,
+                    &mut state.active_query,
+                    state.pass.base.snatch_guard,
+                    state.pass.base.query_set_writes,
+                )
+                .map_pass_err(scope)?;
             }
             ArcComputeCommand::TransitionResources {
                 buffer_transitions,