+++ wgpu-core/src/command/pass.rs
@@ -306,6 +306,8 @@ where
         state.base.raw_encoder,
         query_index,
         pending_query_resets,
+        state.base.snatch_guard,
+        state.base.query_set_writes,
     )?;
     Ok(())
 }